music: split off dates into separate file
Split off Dates and Date Ranges off into their own file.

The Music file was getting too big for it's own good, and the addition
of Date ranges makes splitting it off much easier.
